backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
deepface
numpy
requests
python-dotenv
